fix object picking

+++ Source/Engine/RenderPass/ObjectIDPass.h
@@ -59,7 +59,7 @@ public:
 
 	       int4 ReadBackPixel(const int2& screenCoords) const;
 	inline int4 ReadBackPixel(int screenCoordsX, int screenCoordsY) const { return ReadBackPixel(int2(screenCoordsX, screenCoordsY)); }
-	inline int4 ReadBackPixel(float2 uv) const { return ReadBackPixel(uv.x * mOutputResolutionX, uv.y * mOutputResolutionY); }
+	inline int4 ReadBackPixel(float2 uv) const { return ReadBackPixel((int)uv.x * mOutputResolutionX, (int)uv.y * mOutputResolutionY); }
 	void WaitForCopyComplete() const;
 
 private:
